requirements.txt
absl-py==2.0.0
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
ffmpeg-python==0.2.0
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.42.1
future==0.18.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mediapipe==0.10.5
numpy==1.26.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.1
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.11.2
six==1.16.0
sounddevice==0.4.6
Werkzeug==2.3.7
